Add support for priorityFlashSide option.
This option lets the user specify that a given LED should always be
prioritized when flashing, to make it more noticeable.

This avoids the issue where, for example, if the user is in a meeting
and has another starting within 5 minutes, the flashing may be on
the bottom of the device and more easily missed.

A value of 1 means to prioritize LED1, a value of 2 means to prioritize
LED2.  Any other value is treated as not having priority set.