Major update to add the ability to run calblink as a service.
This also includes one major change and one minor change caused
by this.

Major change: instead of sleeping for 30 seconds (or, during
skip time, up to five minutes), calblink now uses a one-second
timer.  An internal check is made of when the next API call
is needed, so this will reduce the number of API calls made
and increase fidelity in the case where a device is slept.

Minor change: In service mode, the launch header (with the
settings resulting from parsing the config file) will
be skipped and showDots will be forced off.